new file:   APAssignment2/FastReader.java

+++ APAssignment2/Gradable.java
@@ -4,7 +4,7 @@ public interface Gradable {
 
     public void Submit(Student student);
     public void Grade(Instructor instructor);
-    public float viewGrade(Student student);
+    public double viewGrade(Student student);
     public boolean checkStatus();
     public void closeAssessment(Instructor instructor);
     public boolean open();
