+++ APAssignment2/Grade.java
@@ -2,16 +2,16 @@ package APAssignment2;
 
 public class Grade {
     private int max;
-    private float grade;
+    private double grade;
     private Instructor instructor;
 
-    public Grade(int max, float grade, Instructor instructor){
+    public Grade(int max, double grade, Instructor instructor){
         this.max = max;
         this.grade = grade;
         this.instructor = instructor;
     }
 
-    public float getGrade() {
+    public double getGrade() {
         return this.grade;
     }
 }
